+++ fragments/data_vertical.php
@@ -1,0 +1,19 @@
+<div class="row">
+    <div class="col-sm-12">
+        <div class="panel panel-default">
+            <header class="panel-heading">
+                <div class="panel-title"><b><?php echo $this->title; ?></b></div>
+            </header>
+            <div class="panel-body">
+                <div class="row">
+                    <div class="col-sm-12 col-lg-6">
+                        <?php echo $this->chart; ?>
+                    </div>
+                    <div class="col-sm-12 col-lg-6">
+                        <?php echo $this->table; ?>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>