+++ pages/referer.php
@@ -143,7 +143,7 @@ echo $fragment->parse('core/page/section.php');
                 [1, "desc"]
             ],
             "search": {
-                "caseInsensitive": false
+                "caseInsensitive": true
             },
             <?php
 
@@ -188,7 +188,7 @@ echo $fragment->parse('core/page/section.php');
                 [0, "desc"]
             ],
             "search": {
-                "caseInsensitive": false
+                "caseInsensitive": true
             },
             <?php